Elasticsearch 7.x compatibility & cleanups
- Remove legacy SearchMixin and ES classes (they were obsolete,
  everything in addons-server should be using elasticsearch-dsl now)
- Remove 'standard' no-op filter
- Remove doc types, keep backwards-compatible `_doc` for now,
  passing include_type_name=True to support that
